// tests/fixtures/schema-app/apiSchemas.generated.ts
export const apiSchemas = {
// ...
    "stores.search": {
        "input": {
            "type": "object",
            "properties": {
                "query": {
                    "anyOf": [
                        {
                            "type": "string",
                            "minLength": 2
                        },
                        {
                            "type": "object",
                            "properties": {
                                "near": {
                                    "type": "object",
                                    "properties": {
                                        "lat": {
                                            "type": "number",
                                            "minimum": -90,
                                            "maximum": 90
                                        },
                                        "lng": {
                                            "type": "number",
                                            "minimum": -180,
                                            "maximum": 180
                                        }
                                    },
                                    "required": [
                                        "lat",
                                        "lng"
                                    ],
                                    "x-lambder-strip-unknown-keys": true
                                }
                            },
                            "required": [
                                "near"
                            ],
                            "x-lambder-strip-unknown-keys": true
                        }
                    ]
                },
                "limit": {
                    "default": 10,
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 50
                },
                "openNow": {
                    "type": "boolean"
                }
            },
            "required": [
                "query"
            ],
            "x-lambder-strip-unknown-keys": true
        },
        "output": {
            "type": "object",
            "properties": {
                "query": {
                    "anyOf": [
                        {
                            "type": "string"
                        },
                        {
                            "type": "object",
                            "properties": {
                                "near": {
                                    "type": "object",
                                    "properties": {
                                        "lat": {
                                            "type": "number"
                                        },
                                        "lng": {
                                            "type": "number"
                                        }
                                    },
                                    "required": [
                                        "lat",
                                        "lng"
                                    ],
                                    "x-lambder-strip-unknown-keys": true
                                }
                            },
                            "required": [
                                "near"
                            ],
                            "x-lambder-strip-unknown-keys": true
                        }
                    ]
                },
                "limit": {
                    "type": "number"
                },
                "stores": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "name": {
                                "type": "string"
                            },
                            "borough": {
                                "type": "string",
                                "enum": [
                                    "Manhattan",
                                    "Brooklyn",
                                    "Queens",
                                    "Bronx",
                                    "Staten Island"
                                ]
                            }
                        },
                        "required": [
                            "name",
                            "borough"
                        ],
                        "x-lambder-strip-unknown-keys": true
                    }
                }
            },
            "required": [
                "query",
                "limit",
                "stores"
            ],
            "x-lambder-strip-unknown-keys": true
        }
    },
// ...
} as const;